Otautz/feature/#4 template signature (#7)
* Added basic mockup for feature. refs #4

* work in progress update

* work in progress commit.

* instance with assigned parameters is printed correctly :)

* restructure file for packaging

* move to new folder name

* make package ready to be installed with pip

* update readme

* closes #4.
