+++ README.md
@@ -2,7 +2,7 @@
 
 An extension for the ([Semantic](https://www.semantic-mediawiki.org/wiki/Semantic_MediaWiki)) [Media Wiki](https://www.mediawiki.org/wiki/MediaWiki). It parses text in the [OTTR template language](https://ottr.xyz/) and produces code in Media Wiki Syntax. Triples are represented as subobjects.
 
-More information about the extension: https://www.mediawiki.org/wiki/Extension:OttrParser.
+More information about the extension **and how to install**: https://www.mediawiki.org/wiki/Extension:OttrParser.
 
 The help page for wiki usage: https://www.mediawiki.org/wiki/Help:Extension:OttrParser.
 
@@ -23,3 +23,6 @@ In the [Settings.py](Settings.py) you can add your namespaces, that are used in 
 
 ## Ideas
 * currently, it does not add new namespace ids to the wiki
+
+
+
