+++ includes/ottrToSmwPython/NoGenerationErrorStrategy.py
