+++ includes/ottrToSmwPython/Settings.py
