+++ includes/ottrToSmwPython/test_examples/ottr_test_strings.txt
